requirements.txt
icrawler
split-folders
tqdm
pillow
tensorflow
langgraph
langchain
langchain-openai
python-dotenv
streamlit
watchdog